+++ apps/ui/src/components/Collection/CollectionItem/index.spec.tsx
@@ -1,5 +1,5 @@
-import type { ReactNode } from 'react'
 import { cleanup, fireEvent, render, screen } from '@testing-library/react'
+import type { ReactNode } from 'react'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { useMediaServerLibraries } from '../../../api/media-server'
 import CollectionItem from './index'
@@ -37,6 +37,33 @@ describe('CollectionItem', () => {
     cleanup()
   })
 
+  it('keeps the shared poster card aspect ratio so collection tiles stay visible', () => {
+    const { container } = render(
+      <CollectionItem
+        collection={
+          {
+            id: 1,
+            title: 'Action',
+            libraryId: 'library-1',
+            description: 'Collection description',
+            isActive: true,
+            type: 'movie',
+            arrAction: 0,
+            media: [],
+            manualCollection: false,
+            manualCollectionName: '',
+            addDate: new Date(),
+            handledMediaAmount: 0,
+            lastDurationInSeconds: 0,
+            keepLogsForMonths: 0,
+          } as any
+        }
+      />,
+    )
+
+    expect(container.querySelector('[class*="pb-[150%]"]')).toBeTruthy()
+  })
+
   it('delegates direct card clicks to detail navigation when a click handler is provided', () => {
     const openDetail = vi.fn()
 
@@ -67,10 +94,9 @@ describe('CollectionItem', () => {
     fireEvent.click(screen.getByText('Action'))
 
     expect(openDetail).toHaveBeenCalledTimes(1)
-    expect(screen.queryByText('Open Collection')).toBe(null)
   })
 
-  it('opens the shared collection modal when no detail handler is provided', () => {
+  it('does not render a collection modal when no detail handler is provided', () => {
     render(
       <CollectionItem
         collection={
@@ -96,7 +122,6 @@ describe('CollectionItem', () => {
 
     fireEvent.click(screen.getByText('Action'))
 
-    expect(screen.getByText('Close')).toBeTruthy()
-    expect(screen.queryByText('Open Collection')).toBe(null)
+    expect(screen.queryByText('Close')).toBe(null)
   })
 })
