+++ apps/ui/src/components/Common/Poster/PosterCard.spec.tsx
@@ -1,10 +1,24 @@
-import { cleanup, fireEvent, render, screen } from '@testing-library/react'
+import {
+  cleanup,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from '@testing-library/react'
 import { afterEach, describe, expect, it, vi } from 'vitest'
+import GetApiHandler from '../../../utils/ApiHandler'
 import PosterCard from './PosterCard'
 
+vi.mock('../../../utils/ApiHandler', () => ({
+  default: vi.fn(),
+}))
+
 describe('PosterCard', () => {
+  const getApiHandlerMock = vi.mocked(GetApiHandler)
+
   afterEach(() => {
     cleanup()
+    getApiHandlerMock.mockReset()
   })
 
   it('uses the shared container and forwards click interactions', () => {
@@ -20,5 +34,41 @@ describe('PosterCard', () => {
 
     expect(handleClick).toHaveBeenCalledTimes(1)
     expect(screen.getByText('Poster content')).toBeTruthy()
+  })
+
+  it('renders a direct image path without fetching metadata', () => {
+    render(
+      <PosterCard
+        imagePath="https://image.example/poster.jpg"
+        mediaType="movie"
+      >
+        {(image) => <div>{image}</div>}
+      </PosterCard>,
+    )
+
+    expect(screen.getByText('https://image.example/poster.jpg')).toBeTruthy()
+    expect(getApiHandlerMock).not.toHaveBeenCalled()
+  })
+
+  it('resolves metadata-backed images through the shared API path', async () => {
+    getApiHandlerMock.mockResolvedValue({
+      url: 'https://image.example/resolved.jpg',
+    })
+
+    render(
+      <PosterCard mediaType="show" providerIds={{ tmdb: ['123'] }}>
+        {(image) => <div>{image ?? 'missing'}</div>}
+      </PosterCard>,
+    )
+
+    expect(getApiHandlerMock).toHaveBeenCalledWith(
+      '/metadata/image/show?tmdbId=123',
+    )
+
+    await waitFor(() => {
+      expect(
+        screen.getByText('https://image.example/resolved.jpg'),
+      ).toBeTruthy()
+    })
   })
 })
